requirements.txt
aniso8601==8.0.0
azure-cognitiveservices-vision-face==0.4.1
azure-common==1.1.25
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
fac==1.0.0
Flask==1.1.2
Flask-API.yandex==0.6.2.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
msrest==0.6.18
numpy==1.19.1
oauthlib==3.1.0
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
